split out loading/decoding code
